requirements.txt
PyQt5 #(For the interface)

exifread #(For metadata)

folium #(For maps)

tensorflow #(For the AI bonus)

reportlab #(For PDF generation)

Pillow #(For image processing)